requirements.txt
requests==2.26.0
pandas==1.3.3
scikit-learn==0.24.2
nltk==3.6.3
ta==0.7.0
groq==0.4.1
python-dotenv==0.19.0
rich==10.12.0
pycoingecko==3.1.0
yfinance==0.2.18
alpha_vantage==2.3.1